Ch15FastApiSQLAlch/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic[email]
python-dotenv